toolexec: rename injectors to processors
