+++ internal/toolexec/examples/list-go-files.go
@@ -11,10 +11,10 @@ func ExampleListGofiles() {
 	args := []string{"/random/compile", "-trimpath", "randompath", "-p", "random", "-o", "/tmp/randomBuild/_pkg_.a", "-importcfg", "/tmp/random/b002/importcfg", "file1.go", "file2.go", "main.go"}
 	cmd, err := proxy.ParseCommand(args)
 	utils.ExitIfError(err)
-	proxy.ProcessCommand(cmd, InjectCompile)
+	proxy.ProcessCommand(cmd, ProcessCompile)
 }
 
-func InjectCompile(cmd *proxy.CompileCommand) {
+func ProcessCompile(cmd *proxy.CompileCommand) {
 	for _, f := range cmd.GoFiles() {
 		log.Println(f)
 	}
